+++ __tests__/app/message-table.test.tsx
@@ -1,0 +1,89 @@
+import { fireEvent, render, screen } from "@testing-library/react";
+import { MessageTable } from "@/app/message-table";
+import { experimentColumns, RecipeOrBranchInfo } from "@/app/columns";
+import { ExperimentFakes } from "@/__tests__/ExperimentFakes.mjs";
+import { NimbusRecipe } from "@/lib/nimbusRecipe";
+
+describe("MessageTable", () => {
+  describe("ExperimentColumns", () => {
+    it("renders the branch description and slug on expanded branch rows", () => {
+      const rawRecipe = ExperimentFakes.recipe("test-recipe");
+      const nimbusRecipe = new NimbusRecipe(rawRecipe);
+      const messageTableData: RecipeOrBranchInfo[] =
+        [nimbusRecipe.getRecipeInfo()];
+      render(
+        <MessageTable columns={experimentColumns} data={messageTableData} />
+      );
+
+      const toggleButton = screen.getByTestId("toggleBranchRowsButton")
+      fireEvent.click(toggleButton)
+
+      expect(screen.getByRole("table")).toBeInTheDocument();
+      expect(screen.getByText("test description")).toBeInTheDocument();
+      expect(screen.getByText("treatment")).toBeInTheDocument();
+    });
+
+    it("expands all experiment rows when clicking the header toggle button", () => {
+      const rawRecipe1 = ExperimentFakes.recipe("test-recipe-1");
+      const rawRecipe2 = ExperimentFakes.recipe("test-recipe-2");
+      const nimbusRecipe1 = new NimbusRecipe(rawRecipe1);
+      const nimbusRecipe2 = new NimbusRecipe(rawRecipe2);
+      const messageTableData: RecipeOrBranchInfo[] = [
+        nimbusRecipe1.getRecipeInfo(),
+        nimbusRecipe2.getRecipeInfo(),
+      ];
+      render(
+        <MessageTable columns={experimentColumns} data={messageTableData} />
+      );
+
+      const toggleButton = screen.getByTestId("toggleAllRowsButton");
+      fireEvent.click(toggleButton);
+
+      expect(screen.getAllByText("control description").length).toBe(2);
+      expect(screen.getAllByText("control").length).toBe(2);
+      expect(screen.getAllByText("test description").length).toBe(2);
+      expect(screen.getAllByText("treatment").length).toBe(2);
+    });
+
+    it("renders collapsed branch rows at start up", () => {
+      const rawRecipe = ExperimentFakes.recipe("test-recipe");
+      const nimbusRecipe = new NimbusRecipe(rawRecipe);
+      const messageTableData: RecipeOrBranchInfo[] = [
+        nimbusRecipe.getRecipeInfo(),
+      ];
+
+      render(
+        <MessageTable columns={experimentColumns} data={messageTableData} />
+      );
+
+      const controlBranchDescription = screen.queryByText(
+        "control description"
+      );
+      expect(controlBranchDescription).not.toBeInTheDocument();
+      const treatmentBranchDescription = screen.queryByText("test description");
+      expect(treatmentBranchDescription).not.toBeInTheDocument();
+    });
+
+    it("collapses branch rows when clicking the experiment toggle button again after expanding the rows", () => {
+      const rawRecipe = ExperimentFakes.recipe("test-recipe");
+      const nimbusRecipe = new NimbusRecipe(rawRecipe);
+      const messageTableData: RecipeOrBranchInfo[] = [
+        nimbusRecipe.getRecipeInfo(),
+      ];
+      render(
+        <MessageTable columns={experimentColumns} data={messageTableData} />
+      );
+
+      const toggleButton = screen.getByTestId("toggleBranchRowsButton");
+      fireEvent.click(toggleButton);
+      fireEvent.click(toggleButton);
+
+      const controlBranchDescription = screen.queryByText(
+        "control description"
+      );
+      expect(controlBranchDescription).not.toBeInTheDocument();
+      const treatmentBranchDescription = screen.queryByText("test description");
+      expect(treatmentBranchDescription).not.toBeInTheDocument();
+    });
+  });
+});