+++ CHANGELOG.md
@@ -1,9 +1,61 @@
+## Friday, May 24th, 2024
+
+### Added
+
+* Added a Messaging Info navigation menu to include links to key messaging resources
+* Added a Help/Feedback button linking to the #skylight-messaging-system slack channel
+
+## Tuesday, May 21st, 2024
+
+### Changed
+
+* Updated experiment names and branch descriptions to hyperlink to their recipe on Experimenter
+* Removed Experimenter link from the right-most column
+
+## Friday, May 17th, 2024
+
+### Added
+
+* Experiment rows now have toggle buttons that will expand/collapse all its branches
+* The experiments table header now has a toggle button that will expand/collapse all of the experiments at once
+
+### Internal
+
+* SSO integration using the nextjs-auth0 package
+* Middleware that conditionally loads the Auth0 wrapper based on an
+  environment variable (true for prod environment and false for dev && preview)
+
+## Thursday, May 16th, 2024
+
+### Added
+
+* Added Rollout messages and displayed them inside a new table
+* Adding an Info icon tooltip with context to Released table
+
+### Changed
+
+* Update branch names to be human-readable
+* Make table titles clearer
+* Change "Screen 0" to "1st screen" for readability
+* Removed currently information-free Topic column from Released table.
+
+### Fixed
+
+* Only link to branch screenshots that actually exist
+
+### Internal
+
+* Started pulling data from Experimenter instead of Remote Settings to
+  get access to more data for the UI.
+* Support for getting configurables from the environment in preparation for
+  Auth0 and Looker integration.
+
 ## Wednesday, April 24th, 2024
 
 ### Added
 
 * Support for getting configurables from the environment
-* Added a sample .env file which can be copied for local development 
+* Added a sample .env file which can be copied for local development
 
 ## Monday, April 8th, 2024